Change rtorrentscript to use rtxmlrpc
Use pyrocore's rtxmlrpc client instead of xmlrpc

The listening port can now be changed even when ENABLE_RPC2=no.

Back in October @binhex implemented dynamic port reload. Unfortunately
the initial version requires XMLRPC to be exposed to the outside world
(i.e., an RPC2 location has to be exposed in NGINX). When ENABLE_RPC2
is disabled the initial version fails to set the new port, thus
rTorrent becomes unconnectable.

rtxmlrpc on the other hand is able to connect directly to rTorrent's
SGCI port (5000) entirely bypassing NGINX. As a bonus the script does
not need to be aware of NGINX auth credentials.

See also: #124